src/open_source_rag/requirements.txt
torch
faiss-cpu
sentence-transformers
streamlit
llama-cpp-python
langchain
langchain-community
langchain-core
chromadb
tiktoken
